pyproject.toml
[project]
name = "slack-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp==3.9.3",
  "aiosignal==1.3.1",
  "annotated-types==0.6.0",
  "anyio==4.3.0",
  "async-timeout==4.0.3",
  "attrs==23.2.0",
  "autoflake==2.3.0",
  "backoff==2.2.1",
  "black==24.2.0",
  "blinker==1.7.0",
  "botocore==1.34.60",
  "certifi==2024.2.2",
  "charset-normalizer==3.3.2",
  "click==8.1.7",
  "coverage==7.4.3",
  "exceptiongroup==1.2.0",
  "frozenlist==1.4.1",
  "gevent==24.2.1",
  "gql[aiohttp]>=3.5.0",
  "greenlet==3.0.3",
  "gunicorn==21.2.0",
  "h11==0.14.0",
  "httpcore==1.0.4",
  "httpx==0.27.0",
  "idna==3.6",
  "importlib-metadata==7.0.2",
  "inflection>=0.5.1",
  "iniconfig==2.0.0",
  "isort==5.13.2",
  "itsdangerous==2.1.2",
  "jinja2==3.1.3",
  "jmespath==1.0.1",
  "markupsafe==2.1.5",
  "multidict==6.0.5",
  "mypy-extensions==1.0.0",
  "packaging==24.0",
  "pathspec==0.12.1",
  "platformdirs==4.2.0",
  "pluggy==1.4.0",
  "pydantic-core==2.16.3",
  "pydantic==2.6.3",
  "pyflakes==3.2.0",
  "pytest-cov==4.1.0",
  "pytest-mock==3.12.0",
  "pytest-socket==0.7.0",
  "pytest-sugar==1.0.0",
  "pytest==8.1.1",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.0.1",
  "requests-toolbelt==1.0.0",
  "requests==2.31.0",
  "ruff==0.3.2",
  "setuptools==75.6.0",
  "sgqlc>=16.3",
  "six==1.16.0",
  "slack-bolt==1.18.1",
  "slack-sdk==3.27.1",
  "sniffio==1.3.1",
  "termcolor==2.4.0",
  "toml==0.10.2",
  "tomli==2.0.1",
  "typing-extensions==4.10.0",
  "urllib3==1.26.18",
  "websockets==11.0.3",
  "werkzeug==3.0.1",
  "yarl==1.9.4",
  "zipp==3.17.0",
  "zope-event==5.0",
  "zope-interface==6.2",
  "llq-sdk>=0.0.2",
]
